filter array values to php compatible data types

+++ src/Pango/PangoLayout.cpp
@@ -71,13 +71,13 @@ Php::Value PangoLayout_::xy_to_index(Php::Parameters &parameters)
 	if (result) {
 
 		// @TODO wanted canonical implementation
-		// where params returned as generated arguments and return Php::Type::True;
+		// where params return as generated parameters 2,3 and method return Php::Type::True;
 		// https://docs.gtk.org/Pango/method.Layout.xy_to_index.html #143
 
 		Php::Array params;
 
-		params["index_"] = index_;
-		params["trailing"] = trailing;
+		params["index_"] = (int) index_;
+		params["trailing"] = (int) trailing;
 
 		return params;
